refactor(P1.2): extract shared utilities from GenerateView
- Extract CursorAwareTextField/CursorAwareTextView to Utils/CursorAwareTextFields.swift
- Extract personNameScore/bestNameFromCustomFields to Utils/NameScoring.swift
- Update CSVService, RecipientsView, GenerateView to use shared NameScoring utility
- Add new files to Xcode project

DRY refactor: removes ~100 lines of duplicate code across 3 files
