+++ .gitignore
@@ -4,6 +4,7 @@
 
 ## User settings
 xcuserdata/
+prd-specs-changes/
 
 ## compatibility with Xcode 8 and earlier (ignoring not required starting Xcode 9)
 *.xcscmblueprint
